removed Auth service, moved ExternalCredentials to connector project
